// src/types/etudesCourse.ts
export type EtudesCourseSearchItem = {
  id: string;
  sectionSlug: string;
  programmeId: string;
  programmeSlug: string;
  programmeDesignation: string;
  matiereDesignation: string;
  matiereReference: string;
  uniteDesignation: string;
  uniteCode: string;
  semestre: string;
  titulaireName: string;
  titulaireEmail: string;
  horaireJour: string;
  horaireHeureDebut: string;
  horaireHeureFin: string;
  status: string;
  searchText: string;
};